+++ src/index.d.ts
@@ -1,0 +1,7 @@
+import useSpring from './hooks/useSpring';
+import useSprings from './hooks/useSprings';
+import useTrail from './hooks/useTrail';
+import Controller from './Controller';
+import constants from './constants';
+
+export { useSpring, useSprings, useTrail, Controller, constants };