+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added `progress` config for easing animations ([@chriscerie](https://github.com/chriscerie) in [#13](https://github.com/chriscerie/roact-spring/pull/13))
 * Hooks now cancel animations when they are unmounted
+* Added staggered text story to demos
 
 ## 0.2.1 (Feburary 17, 2022)